WayneBlog/config/serializers.py
from rest_framework import serializers

from .models import Link


class LinkSerializer(serializers.ModelSerializer):
    class Meta:
        model = Link
        fields = (
            'id', 'title', 'href', 'created_time',
        )
